travis: Set variable for Darwin and check that.
There were a bunch of `$(uname -s) == 'Darwin'` calls which have now
been consolidated into one setting that's being checked everywhere.

Additionally one branch was inverted. We were using `if DARWIN; else`
everywhere except for in one place where the negated form was used. In
order to keep it consistent with the rest of the script the branches
have been swapped.