Prepares X509 decoding functions for more CA functionality.
Most of the existing `_build_*()` methods will need to be reused for CSR
extensions, so it is more convenient to have them as free functinos.

Since we will need to add the reverse functions, it is convenient to
rename the `_build_*()` methods to `_decode_*()`.  The reverse
functions will be named `_encode_*()`.